config: newArg returns error, PubSub filtered by paths parts
type arg contains now a slice and pathLevelX() functions which are
used in the pubsub system to allow subscription to a specific topic.
A topic is a part of a path or full path.
Subscribers which panics will get removed auotmatically without crashing
the system.
The Subscriber interface has been renamed because in other packages
there might exists other MessageX() functions for other pubsubs.