net/internal/scopedservice: Refactor how config gets inherited from parents
WIP

Move: TypeID as variadic arguments to the end so you can specify a
hierarchy. The first argument of the variadic TypeID can be left off
then a fall back to scope Default would happen. If provided the
configuration gets assigned to that scope. Subsequent scopes are treated
as inheritances.

All functional options must be refactored.