net/signed/backend: Adjust to new API of net/internal/scopedservice
Also fixing this bug:

i think there is a bug with the scpHash because scpHash returns that
hash in which scope the value has been found. for example we have 2
websites with each 2 stores. those 4 stores are configured for
signing. so we have for each store scope in the scopeCache map an
entry. if we now disable the signing for a website then it won't be
disabled for the store scopes because it won't fall back to website
as we have found an entry in the store scope in the ScopeCache map.
hence the solution would be to use the hash value from sg.Scope()