storage/dbr: Add Type Base
Base represents the commonly used fields for each struct for a database table
or a view. Base scans a *sql.Rows into a *sql.RawBytes slice and allows to
convert the byte slices into the desired type without allocating memory. Base
should be used as a composite field in a database table struct.