storage/dbr: Big refactor of WhereFragment to rename to Condition
A Condition implements a type safe way to add arguments. This is the
first step to get rid of all the exported Argument-implementing types.
A Condition can not only be part of a WHERE/JOIN/HAVING statement even
in the SET claus of an UPDATE statement. So Condition types are more
reuseable.