storage/dbr: Refactor Binder interface, add Bind* functions to DML
SQLStmt got removed because no need to know in which statement
the Binder gets called.
Each of the four Statements got the function Bind and BindByQualifier
to assign a Binder to either a table/view/etc name or its alias.
Reason to keep the Binder.AppendBind function free of used aliases
in the statement. The AppendBind does not need to know about the internals
of a SQL query.