storage/dbr: Introduced qualified records, Insert/Delete Prepared
A type can implement ArgumentsAppender and hence it is called a record.
When a record gets assigned to a SQL statement you must bind it to a qualifier
otherwise that record does not now to which table/view or alias it
belongs to.

Refactored also the Binder interface as Binder is the wrong name.

Insert and Delete became new Prepare functions which return a dedicated
type for running the prepared staement with different arguments in a
loop.