sql/dml: Add BenchmarkSelectExcessConditions
Type Condition is not a pointer. The GC pressure is much less but he program
is a bit slower due to copying.
BenchmarkSelectExcessConditions-4   	  200000	      7055 ns/op	    4984 B/op	      26 allocs/op
BenchmarkSelectExcessConditions-4   	  200000	      7118 ns/op	    4984 B/op	      26 allocs/op
The next line gives the results for the implementation when using type
Condition as a pointer. We have more allocs, more pressure on the GC but it's
a bit faster.
BenchmarkSelectExcessConditions-4   	  200000	      6297 ns/op	    4920 B/op	      35 allocs/op
For now we stick with the pointers. Reason: the SQL statement gets only
created once and not in a loop.