config: Enable hot reloading via OS signal and environment awareness
An OS signal like USR1/2 etc can be send to the Go program to force
reloading of the provided config.LoadDataFn functions.

The name of the OS environment variable DefaultOSEnvVariableName
gets read and the value gets used as the current environment in which
this service runs. For example: STAGING, PROUDCTION, CI.
With this information the package storage/cfgfile can load files depending
on that environment to provide different configuration values.

Type Path is not yet aware of the environment.