sql/dmlgen: Add Copy and WriteTo, feature switch, custom code
The feature switch allows to enable or disable collection/entity methods.

WriteTo function writes the struct fields and data to the io.Writer.
This is useful for generating a hash of a struct.

The custom code functionality allows the user to swap out default
implementations and replace them with their own.