sql/dml: Tuples in WHERE clause
Allows to use tuples in a WHERE clause.

    SELECT * FROM `core_config_data` WHERE
    ((`entity_id`,`attribute_id`,`store_id`) IN (?,?,?))

The raw SQL query contains a tuples placeholder because the amount
of arguments might vary.

     WHERE ((`entity_id`, `attribute_id`, `store_id`, `source_id`)
     IN /*TUPLES=004*/)

The 004 indicates that 4 columns are used so 4 placeholders are needed.
The overall code requires some refactoring but so far it works
performant.

When running DBR, with the option ExpandPlaceholders, the tuple
placeholder will be expanded to the correct amount of arguments.

    SELECT * FROM `core_config_data` WHERE
    ((`entity_id`,`attribute_id`,`store_id`) IN (?,?,?),(?,?,?))

Only DELETE, UPDATE and SELECT support the tuples.