sql/dml: With* improvements
withExecSQL runs only SQL queries in a transaction when there are more
than one queries supplied. Reason: queries like CREATE DATABASE
can't run in a transaction.

WithDSN can create and drop a temporary database for testing purposes
when the database name in the DSN is equal to the word "random".